Fix buggy mime-type detection
The buffer mime-type detection was assuming we'd only get the first few
bytes of a file because at one point that is how it was being called.
Some callers may want to send more content for better magic detection
but this causes issues with the edifact/edi over-rides. Changing the
logic to only read the first few bytes for edifact/edi detection.