More consistant dfs cache updates
Do not rely on the timestamp we get from the dfs cache service since it
can get stale. This causes the consumers to check for a new cache on
every request until the timestamp gets updated. Instead, the client will
update it's own timestamp so that it will fetch a new cache at a
consistan interval of 5 minutes and will check at most once every 5
minutes.